backend/requirements.txt
Flask
flask-cors
flask_sqlalchemy
pdfminer.six
python-docx
sentence-transformers  # optional - faster/better matching if installed